# loop_implementation/api/openapi.yaml
openapi: 3.1.0
info:
  title: Loop Platform API
  version: 0.1.0
  description: |
    Public REST API for the Loop agent platform.

    All endpoints require a Bearer token scoped to a workspace.
    Streaming endpoints additionally support Server-Sent Events (SSE) and WebSocket.
  contact:
    name: Loop Engineering
    email: eng@loop.example
  license:
    name: Apache-2.0
servers:
  - url: https://api.loop.example/v1
    description: Production (na-east — abstract region; concrete cloud may be AWS/Azure/GCP/Alibaba/self-host)
  - url: https://api-eu.loop.example/v1
    description: Production (eu-west — same)
  - url: http://localhost:8080/v1
    description: Local dev
security:
  - BearerAuth: []
...
paths:
...
  /workspaces:
    get:
      tags: [Workspaces]
      summary: List workspaces the caller is a member of
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceList"
    post:
      tags: [Workspaces]
      summary: Create workspace
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspaceCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
  /workspaces/{workspace_id}:
    get:
      tags: [Workspaces]
      summary: Get workspace
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/RequestRegion" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Workspaces]
      summary: Update workspace metadata
      description: Workspace `region` is immutable after creation.
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/RequestRegion" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspacePatch" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /workspaces/{workspace_id}/cobuilder:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [CoBuilder]
      summary: Get AI Co-Builder suggestions for workspace state
      parameters:
        - in: query
          name: agent_id
          schema: { type: string, format: uuid }
          required: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CoBuilderWorkspace" }
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PASETO
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
...
  parameters:
    WorkspaceId:
      in: path
      name: workspace_id
      required: true
      schema: { type: string, format: uuid }
...
    RequestRegion:
      in: header
      name: X-Loop-Region
      required: false
      description: Abstract region serving the request; mismatches with workspace.region return 403.
      schema: { $ref: "#/components/schemas/RegionName" }
...
  schemas:
...
    ProblemDetail:
      description: RFC 9457 Problem Details for HTTP APIs
      type: object
      required: [type, status, title]
      properties:
        type: { type: string, format: uri, example: "https://api.loop.example/errors/rate-limit" }
        status: { type: integer, example: 429 }
        title: { type: string, example: "Too Many Requests" }
        detail: { type: string }
        instance: { type: string, format: uri }
        request_id: { type: string, description: "Correlation ID for logs" }
    Workspace:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        plan: { type: string, enum: [hobby, pro, team, enterprise] }
        region: { $ref: "#/components/schemas/RegionName" }
        created_at: { type: string, format: date-time }
        role: { type: string, enum: [owner, admin, member, viewer] }
    WorkspaceList:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Workspace" }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
    WorkspaceCreate:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        slug: { type: string, pattern: "^[a-z0-9][a-z0-9-]*$" }
        region:
          allOf: [{ $ref: "#/components/schemas/RegionName" }]
          default: na-east
          description: Immutable workspace region pin selected at creation.
    WorkspacePatch:
      type: object
      description: Mutable workspace metadata. Region changes are rejected.
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
    RegionName:
      type: string
      enum: [na-east, eu-west]
      description: Abstract Loop region from infra/terraform/regions.yaml.
...
    OperatorContext:
      type: object
      required: [maxMode, scopes, budgetRemainingUsd]
      properties:
        maxMode: { type: string, enum: [suggest, edit, drive] }
        scopes:
          type: array
          items: { type: string }
        budgetRemainingUsd: { type: number }
    CoBuilderProvenance:
      type: object
      required: [source, excerpt, evidenceRef]
      properties:
        source: { type: string }
        excerpt: { type: string }
        evidenceRef: { type: string }
    CoBuilderCost:
      type: object
      required: [usd, latencyMs]
      properties:
        usd: { type: number }
        latencyMs: { type: integer }
    CoBuilderDiffHunk:
      type: object
      required: [path, body]
      properties:
        path: { type: string }
        body: { type: string }
    CoBuilderAction:
      type: object
      required:
        - id
        - mode
        - title
        - rationale
        - diff
        - provenance
        - cost
        - requiredScopes
        - evidenceRef
      properties:
        id: { type: string }
        mode: { type: string, enum: [suggest, edit, drive] }
        title: { type: string }
        rationale: { type: string }
        diff: { $ref: "#/components/schemas/CoBuilderDiffHunk" }
        provenance:
          type: array
          items: { $ref: "#/components/schemas/CoBuilderProvenance" }
        cost: { $ref: "#/components/schemas/CoBuilderCost" }
        requiredScopes:
          type: array
          items: { type: string }
        evidenceRef: { type: string }
    RubberDuckFinding:
      type: object
      required: [step, observation, evidenceRef]
      properties:
        step: { type: string }
        observation: { type: string }
        evidenceRef: { type: string }
    RubberDuckDiagnosis:
      type: object
      required: [caseId, failureSummary, findings, proposedFix]
      properties:
        caseId: { type: string }
        failureSummary: { type: string }
        findings:
          type: array
          items: { $ref: "#/components/schemas/RubberDuckFinding" }
        proposedFix: { $ref: "#/components/schemas/CoBuilderAction" }
    ReviewBullet:
      type: object
      required: [id, severity, body, evidenceRef]
      properties:
        id: { type: string }
        severity: { type: string, enum: [info, warn, block] }
        body: { type: string }
        evidenceRef: { type: string }
    AdversarialReview:
      type: object
      required: [actionId, bullets]
      properties:
        actionId: { type: string }
        bullets:
          type: array
          items: { $ref: "#/components/schemas/ReviewBullet" }
    CoBuilderWorkspace:
      type: object
      required: [workspaceId, agentId, agentName, operator, actions, rubberDuck, review]
      properties:
        workspaceId: { type: string, format: uuid }
        agentId: { type: string, format: uuid, nullable: true }
        agentName: { type: string }
        operator: { $ref: "#/components/schemas/OperatorContext" }
        actions:
          type: array
          items: { $ref: "#/components/schemas/CoBuilderAction" }
        rubberDuck: { $ref: "#/components/schemas/RubberDuckDiagnosis" }
        review: { $ref: "#/components/schemas/AdversarialReview" }
...